fix some compiler warnings
* delete called on 'AudioData' that is abstract but has non-virtual destructor
* '~RandomSource' has a non-throwing exception specification but can still throw
* result of comparison of constant 18446744073709551615 with expression of type 'unsigned int' is always false
* non-void function does not return a value

+++ src/Utils.cc
@@ -45,7 +45,7 @@ std::string Utils::formatHRSize(unsigned long size) {
 }
 
 std::string Utils::stripDir(std::string s) {
-    unsigned int start = 0;
+    size_t start = 0;
     if ((start = s.find_last_of("/\\")) == std::string::npos) {
         start = 0;
     } else {
